refactor: вынес логику определения разрешения экрана и смены названия страниц в кастомные хуки
