+++ src/pages/FullNews.tsx
@@ -11,11 +11,11 @@ const FullNews = () => {
   const dispatch = useAppDispatch();
   const { item, status } = useSelector(selectNewsData);
 
-  console.log(item);
+  // console.log(item);
 
   const [isSharePanelOpen, setIsSharePanelOpen] = React.useState<boolean>(false);
   const { id } = useParams();
-  const navigate = useNavigate();
+  // const navigate = useNavigate();
 
   React.useEffect(() => {
     if (id) {
