+++ src/pages/FullProject.tsx
@@ -13,7 +13,7 @@ const FullProject = () => {
   const [isSharePanelOpen, setIsSharePanelOpen] = React.useState<boolean>(false);
 
   const { id } = useParams();
-  const navigate = useNavigate();
+  // const navigate = useNavigate();
 
   React.useEffect(() => {
     if (id) {
