+++ src/pages/News.tsx
@@ -6,41 +6,15 @@ import { useSelector } from 'react-redux';
 import { selectNewsData } from '../redux/news/selectors';
 import { setCurrentPage } from '../redux/news/slice';
 import { menuItems } from '../utils/constants';
+import { useResize } from '../hooks/useResize';
 
 const News = () => {
   const dispatch = useAppDispatch();
-
   const { items, status, currentPage, totalPages } = useSelector(selectNewsData);
 
+  const { screenWidth } = useResize();
+
   const [fetching, setFetching] = React.useState(false);
-  const [screenWidth, setScreenWidth] = React.useState(window.innerWidth);
-
-  React.useEffect(() => {
-    document.title = 'Новости';
-  }, []);
-
-  React.useEffect(() => {
-    const handleResize = () => {
-      setScreenWidth(window.innerWidth);
-    };
-
-    let timeoutId: NodeJS.Timeout;
-
-    const delayedHandleResize = () => {
-      clearTimeout(timeoutId);
-
-      timeoutId = setTimeout(() => {
-        handleResize();
-      }, 500);
-    };
-
-    window.addEventListener('resize', delayedHandleResize);
-
-    return () => {
-      window.removeEventListener('resize', delayedHandleResize);
-      clearTimeout(timeoutId);
-    };
-  }, []);
 
   React.useEffect(() => {
     const handleSetLimit = () => {
