+++ src/pages/Reports.tsx
@@ -6,42 +6,18 @@ import { useSelector } from 'react-redux';
 import { selectReportData } from '../redux/report/selectors';
 import { fetchReport } from '../redux/report/asyncActions';
 import { menuItems } from '../utils/constants';
+import { useResize } from '../hooks/useResize';
 
 const Reports = () => {
   const dispatch = useAppDispatch();
   const { items, status } = useSelector(selectReportData);
+
+  const { screenWidth } = useResize();
+
   const [reportIndex, setReportIndex] = React.useState<number>(0);
   const [isReportPopupOpen, setReportPopupOpen] = React.useState<boolean>(false);
-  const [screenWidth, setScreenWidth] = React.useState<number>(window.innerWidth);
 
   const isMounted = React.useRef<boolean>(false);
-
-  React.useEffect(() => {
-    document.title = 'Отчёты';
-  }, []);
-
-  React.useEffect(() => {
-    const handleResize = () => {
-      setScreenWidth(window.innerWidth);
-    };
-
-    let timeoutId: NodeJS.Timeout;
-
-    const delayedHandleResize = () => {
-      clearTimeout(timeoutId);
-
-      timeoutId = setTimeout(() => {
-        handleResize();
-      }, 300);
-    };
-
-    window.addEventListener('resize', delayedHandleResize);
-
-    return () => {
-      window.removeEventListener('resize', delayedHandleResize);
-      clearTimeout(timeoutId);
-    };
-  }, []);
 
   React.useEffect(() => {
     if (!isMounted.current) {
