+++ src/pages/UnionMembers.tsx
@@ -8,44 +8,17 @@ import { useAppDispatch } from '../redux/store';
 import clsx from 'clsx';
 import { fetchUnionMembers } from '../redux/unionMember/asyncActions';
 import { menuItems } from '../utils/constants';
+import { useResize } from '../hooks/useResize';
 
 const UnionMembers: React.FC = () => {
   const dispatch = useAppDispatch();
 
   const { items, currentPage, totalPages, status } = useSelector(selectUnionMembersData);
 
-  const [screenWidth, setScreenWidth] = React.useState<number>(window.innerWidth);
-  const [clientWidth, setClientWidth] = React.useState<number>(document.documentElement.clientWidth);
+  const { screenWidth, clientWidth } = useResize();
+
   const [cardsLimit, setCardsLimit] = React.useState<number>(0);
   const [fetching, setFetching] = React.useState(false);
-
-  React.useEffect(() => {
-    document.title = 'Состав';
-  }, []);
-
-  React.useEffect(() => {
-    const handleResize = () => {
-      setScreenWidth(window.innerWidth);
-      setClientWidth(document.documentElement.clientWidth);
-    };
-
-    let timeoutId: NodeJS.Timeout;
-
-    const delayedHandleResize = () => {
-      clearTimeout(timeoutId);
-
-      timeoutId = setTimeout(() => {
-        handleResize();
-      }, 500);
-    };
-
-    window.addEventListener('resize', delayedHandleResize);
-
-    return () => {
-      window.removeEventListener('resize', delayedHandleResize);
-      clearTimeout(timeoutId);
-    };
-  }, []);
 
   React.useEffect(() => {
     const handleSetLimit = (): number => {
